+++ src/coordinator/roles.spec.js
@@ -29,6 +29,207 @@ describe('Coordinator/Roles', function() {
     this.sandbox.restore();
   });
 
+  describe('addApplication', function() {
+    context('when all the required parameters are provided', function() {
+      let expectedRoleApplication;
+      let promise;
+      let request;
+      let application;
+      let role;
+      let roleApplicationFromServer;
+      let toCamelCase;
+
+      beforeEach(function() {
+        application = fixture.build('contxtApplication');
+        role = fixture.build('contxtRole');
+
+        expectedRoleApplication = fixture.build('contxtRoleApplication');
+        roleApplicationFromServer = fixture.build(
+          'contxtRoleApplication',
+          expectedRoleApplication,
+          {
+            fromServer: true
+          }
+        );
+
+        request = {
+          ...baseRequest,
+          post: this.sandbox.stub().resolves(roleApplicationFromServer)
+        };
+        toCamelCase = this.sandbox
+          .stub(objectUtils, 'toCamelCase')
+          .callsFake((app) => expectedRoleApplication);
+
+        const roles = new Roles(baseSdk, request, expectedHost);
+        promise = roles.addApplication(role.id, application.id);
+      });
+
+      it('posts the role application to the server', function() {
+        expect(request.post).to.be.calledWith(
+          `${expectedHost}/applications/${application.id}/roles/${role.id}`
+        );
+      });
+
+      it('formats the returned role application', function() {
+        return promise.then(() => {
+          expect(toCamelCase).to.be.calledWith(roleApplicationFromServer);
+        });
+      });
+
+      it('returns a fulfilled promise', function() {
+        return expect(promise).to.be.fulfilled.and.to.eventually.deep.equal(
+          expectedRoleApplication
+        );
+      });
+    });
+
+    context('when the role ID is not provided', function() {
+      it('throws an error', function() {
+        const application = fixture.build('contxtApplication');
+
+        const roles = new Roles(baseSdk, baseRequest, expectedHost);
+        const promise = roles.addApplication(null, application.id);
+
+        return expect(promise).to.be.rejectedWith(
+          'A roleId is required for adding an application to a role'
+        );
+      });
+    });
+
+    context('when the application ID is not provided', function() {
+      it('throws an error', function() {
+        const role = fixture.build('contxtRole');
+
+        const roles = new Roles(baseSdk, baseRequest, expectedHost);
+        const promise = roles.addApplication(role.id, null);
+
+        return expect(promise).to.be.rejectedWith(
+          'An applicationId is required for adding an application to a role'
+        );
+      });
+    });
+  });
+
+  describe('addStack', function() {
+    context(
+      'when all the required parameters are provided and valid',
+      function() {
+        let expectedRoleStack;
+        let promise;
+        let request;
+        let stack;
+        let role;
+        let roleStackFromServer;
+        let toCamelCase;
+
+        beforeEach(function() {
+          stack = fixture.build('contxtStack');
+          role = fixture.build('contxtRole');
+
+          expectedRoleStack = fixture.build('contxtRoleStack');
+          roleStackFromServer = fixture.build(
+            'contxtRoleStack',
+            expectedRoleStack,
+            {
+              fromServer: true
+            }
+          );
+
+          request = {
+            ...baseRequest,
+            post: this.sandbox.stub().resolves(roleStackFromServer)
+          };
+          toCamelCase = this.sandbox
+            .stub(objectUtils, 'toCamelCase')
+            .callsFake((app) => expectedRoleStack);
+
+          const roles = new Roles(baseSdk, request, expectedHost);
+          promise = roles.addStack(
+            role.id,
+            stack.id,
+            expectedRoleStack.accessType
+          );
+        });
+
+        it('posts the role stack to the server', function() {
+          expect(request.post).to.be.calledWith(
+            `${expectedHost}/stacks/${stack.id}/roles/${role.id}`,
+            {
+              access_type: expectedRoleStack.accessType
+            }
+          );
+        });
+
+        it('formats the returned role stack', function() {
+          return promise.then(() => {
+            expect(toCamelCase).to.be.calledWith(roleStackFromServer);
+          });
+        });
+
+        it('returns a fulfilled promise', function() {
+          return expect(promise).to.be.fulfilled.and.to.eventually.deep.equal(
+            expectedRoleStack
+          );
+        });
+      }
+    );
+
+    context('when the role ID is not provided', function() {
+      it('throws an error', function() {
+        const stack = fixture.build('contxtStack');
+        const roleStack = fixture.build('contxtRoleStack');
+
+        const roles = new Roles(baseSdk, baseRequest, expectedHost);
+        const promise = roles.addStack(null, stack.id, roleStack.accessType);
+
+        return expect(promise).to.be.rejectedWith(
+          'A roleId is required for adding a stack to a role.'
+        );
+      });
+    });
+
+    context('when the stack ID is not provided', function() {
+      it('throws an error', function() {
+        const role = fixture.build('contxtRole');
+        const roleStack = fixture.build('contxtRoleStack');
+        const roles = new Roles(baseSdk, baseRequest, expectedHost);
+        const promise = roles.addStack(role.id, null, roleStack.accessType);
+
+        return expect(promise).to.be.rejectedWith(
+          'A stackId is required for adding a stack to a role.'
+        );
+      });
+    });
+
+    context('when the access type is not provided', function() {
+      it('throws an error', function() {
+        const role = fixture.build('contxtRole');
+        const stack = fixture.build('contxtStack');
+
+        const roles = new Roles(baseSdk, baseRequest, expectedHost);
+        const promise = roles.addStack(role.id, stack.id, null);
+
+        return expect(promise).to.be.rejectedWith(
+          'An accessType of "reader", "collaborator", or "owner" is required for adding a stack to a role.'
+        );
+      });
+    });
+
+    context('when the access type is not a valid value', function() {
+      it('throws an error', function() {
+        const role = fixture.build('contxtRole');
+        const stack = fixture.build('contxtStack');
+
+        const roles = new Roles(baseSdk, baseRequest, expectedHost);
+        const promise = roles.addStack(role.id, stack.id, faker.random.word());
+
+        return expect(promise).to.be.rejectedWith(
+          'An accessType of "reader", "collaborator", or "owner" is required for adding a stack to a role.'
+        );
+      });
+    });
+  });
+
   describe('constructor', function() {
     let roles;
 
@@ -270,6 +471,106 @@ describe('Coordinator/Roles', function() {
 
         return expect(promise).to.be.rejectedWith(
           'An organizationId is required for getting roles for an organization.'
+        );
+      });
+    });
+  });
+
+  describe('removeApplication', function() {
+    context('when all required parameters are provided', function() {
+      let application;
+      let role;
+      let promise;
+
+      beforeEach(function() {
+        application = fixture.build('contxtApplication');
+        role = fixture.build('contxtRole');
+
+        const roles = new Roles(baseSdk, baseRequest, expectedHost);
+        promise = roles.removeApplication(role.id, application.id);
+      });
+
+      it('sends a request to removeApplication from the role', function() {
+        expect(baseRequest.delete).to.be.calledWith(
+          `${expectedHost}/applications/${application.id}/roles/${role.id}`
+        );
+      });
+
+      it('returns a resolved promise', function() {
+        return expect(promise).to.be.fulfilled;
+      });
+    });
+
+    context('when the role ID is not provided', function() {
+      it('throws an error', function() {
+        const application = fixture.build('contxtApplication');
+        const roles = new Roles(baseSdk, baseRequest, expectedHost);
+        const promise = roles.removeApplication(null, application.id);
+
+        return expect(promise).to.be.rejectedWith(
+          'A roleId is required for removing an application from a role.'
+        );
+      });
+    });
+
+    context('when the application ID is not provided', function() {
+      it('throws an error', function() {
+        const role = fixture.build('contxtRole');
+        const roles = new Roles(baseSdk, baseRequest, expectedHost);
+        const promise = roles.removeApplication(role.id, null);
+
+        return expect(promise).to.be.rejectedWith(
+          'An applicationId is required for removing an application from a role.'
+        );
+      });
+    });
+  });
+
+  describe('removeStack', function() {
+    context('when all required parameters are provided', function() {
+      let stack;
+      let role;
+      let promise;
+
+      beforeEach(function() {
+        stack = fixture.build('contxtStack');
+        role = fixture.build('contxtRole');
+
+        const roles = new Roles(baseSdk, baseRequest, expectedHost);
+        promise = roles.removeStack(role.id, stack.id);
+      });
+
+      it('sends a request to removeStack the role from the organization', function() {
+        expect(baseRequest.delete).to.be.calledWith(
+          `${expectedHost}/stacks/${stack.id}/roles/${role.id}`
+        );
+      });
+
+      it('returns a resolved promise', function() {
+        return expect(promise).to.be.fulfilled;
+      });
+    });
+
+    context('when the role ID is not provided', function() {
+      it('throws an error', function() {
+        const stack = fixture.build('contxtStack');
+        const roles = new Roles(baseSdk, baseRequest, expectedHost);
+        const promise = roles.removeStack(null, stack.id);
+
+        return expect(promise).to.be.rejectedWith(
+          'A roleId is required for removing a stack from a role.'
+        );
+      });
+    });
+
+    context('when the stack ID is not provided', function() {
+      it('throws an error', function() {
+        const role = fixture.build('contxtRole');
+        const roles = new Roles(baseSdk, baseRequest, expectedHost);
+        const promise = roles.removeStack(role.id, null);
+
+        return expect(promise).to.be.rejectedWith(
+          'A stackId is required for removing a stack from a role.'
         );
       });
     });